+++ test/e2e/serial/tests/resource_accounting.go
@@ -156,7 +156,7 @@ var _ = Describe("[serial][disruptive][scheduler][resacct] numaresources workloa
 			}
 
 			By("Waiting for padding pods to be ready")
-			failedPodIds := e2ewait.ForPaddingPodsRunning(fxt, paddingPods)
+			failedPodIds := e2efixture.WaitForPaddingPodsRunning(fxt, paddingPods)
 			Expect(failedPodIds).To(BeEmpty(), "some padding pods have failed to run")
 
 			var targetNrtBefore *nrtv1alpha1.NodeResourceTopology
@@ -189,7 +189,7 @@ var _ = Describe("[serial][disruptive][scheduler][resacct] numaresources workloa
 			}
 
 			By("Waiting for padding pods to be ready")
-			failedPodIds = e2ewait.ForPaddingPodsRunning(fxt, targetPaddingPods)
+			failedPodIds = e2efixture.WaitForPaddingPodsRunning(fxt, targetPaddingPods)
 			Expect(failedPodIds).To(BeEmpty(), "some padding pods have failed to run")
 
 			By("saturating nodes we want to be unsuitable")
@@ -216,7 +216,7 @@ var _ = Describe("[serial][disruptive][scheduler][resacct] numaresources workloa
 			allPaddingPods = append(allPaddingPods, targetPaddingPods...)
 
 			By("Waiting for padding pods to be ready")
-			failedPodIds = e2ewait.ForPaddingPodsRunning(fxt, allPaddingPods)
+			failedPodIds = e2efixture.WaitForPaddingPodsRunning(fxt, allPaddingPods)
 			Expect(failedPodIds).To(BeEmpty(), "some padding pods have failed to run")
 
 			// TODO: smarter cooldown
@@ -382,7 +382,7 @@ var _ = Describe("[serial][disruptive][scheduler][resacct] numaresources workloa
 			}
 
 			By("Waiting for padding pods to be ready")
-			failedPodIds := e2ewait.ForPaddingPodsRunning(fxt, paddingPods)
+			failedPodIds := e2efixture.WaitForPaddingPodsRunning(fxt, paddingPods)
 			Expect(failedPodIds).To(BeEmpty(), "some padding pods have failed to run")
 		})
