+++ test/e2e/serial/tests/workload_overhead.go
@@ -208,7 +208,7 @@ var _ = Describe("[serial][disruptive][scheduler] numaresources workload overhea
 				}
 
 				By("Waiting for padding pods to be ready")
-				failedPodIds := e2ewait.ForPaddingPodsRunning(fxt, paddingPods)
+				failedPodIds := e2efixture.WaitForPaddingPodsRunning(fxt, paddingPods)
 				Expect(failedPodIds).To(BeEmpty(), "some padding pods have failed to run")
 
 				By("checking the resource allocation as the test starts")
